+++ source/libraries/mn-toolkit/library/components/container/Grid.tsx
@@ -1,8 +1,6 @@
-import { ReactElement, ReactNode } from 'react';
-import { TJSXElementChildren } from '../../system';
 import { IContainerProps, IContainerState, Container } from './Container';
 
-export interface IGridProps extends IContainerProps {}
+export interface IGridProps extends IContainerProps<HTMLDivElement, 'grid'> {}
 
 export interface IGridState extends IContainerState {}
 
@@ -12,29 +10,8 @@ export class Grid<P extends IGridProps = IGridProps, S extends IGridState = IGri
       ...super.defaultProps,
       layout: 'grid',
       gutter: true,
-      wrap: true,
+      gridColumns: 12,
+      gridRows: 1,
     };
-  }
-
-  public override get children(): TJSXElementChildren {
-    return (this.props.children as unknown as ReactNode[])
-      .filter((x) => !!x)
-      .map((x, i) => (
-        <div key={`mn-grid-item-${i}`} className={this.getGridItemClasses(x as ReactElement)}>
-          {x}
-        </div>
-      ));
-  }
-
-  protected getGridItemClasses(x: ReactElement): string {
-    const classes = ['mn-grid-item'];
-    if (!!x.props) {
-      if (!!x.props.s) classes.push(`mn-containable-item-width-s-${x.props.s}`);
-      if (!!x.props.m) classes.push(`mn-containable-item-width-m-${x.props.m}`);
-      if (!!x.props.l) classes.push(`mn-containable-item-width-l-${x.props.l}`);
-      if (!!x.props.xl) classes.push(`mn-containable-item-width-xl-${x.props.xl}`);
-      if (!!x.props.xxl) classes.push(`mn-containable-item-width-xxl-${x.props.xxl}`);
-    }
-    return classes.join(' ');
   }
 }