+++ source/libraries/mn-toolkit/library/components/container/HorizontalStack.tsx
@@ -1,6 +1,7 @@
 import { IContainerProps, IContainerState, Container } from './Container';
 
 export interface IHorizontalStackProps extends IContainerProps {}
+
 interface IHorizontalStackState extends IContainerState {}
 
 export class HorizontalStack extends Container<IHorizontalStackProps, IHorizontalStackState> {