+++ source/libraries/mn-toolkit/library/components/dateTime/choosers/DateChooser.tsx
@@ -235,9 +235,10 @@ export class DateChooser extends Container<IDateChooserProps, IDateChooserState>
       years.push(
         <HorizontalStack
           key={year}
-          s='4'
-          m='4'
-          l='3'
+          colSpans={{
+            small: 4,
+            large: 3,
+          }}
           fill
           className={classNames('year', { selected: year === selectedYear })}
           itemAlignment='center'